# documentazione e database/openapi.yaml
openapi: 3.0.3
info:
  title: Fooody REST API
  description: API RESTful per la gestione delle ordinazioni e delle consegne a domicilio dell'attività Fooody.
  version: 1.0.0
  contact:
    name: Maikol Gaparroni
    email: maikol.gasparroni@student.univaq.it
servers:
  - url: http://localhost:8080/Fooody/api
    description: Server locale (Apache Tomcat)
paths:
...
  /prodotti:
    get:
      summary: Elenco di tutti i prodotti e ricerca con filtri.
      description: Restituisce una lista di tutti i prodotti disponibili nel sistema. Supporta la ricerca con filtri per nome, fascia di prezzo e categoria.
      tags: [Catalogo]
      parameters:
        - in: query
          name: categoria
          schema:
            type: string
          description: Filtro per categoria del prodotto (es. "Pizza", "Bevande").
        - in: query
          name: nome
          schema:
            type: string
          description: Filtro per nome del prodotto.
        - in: query
          name: prezzoMin
          schema:
            type: number
            format: float
          description: Filtro per prezzo minimo del prodotto.
        - in: query
          name: prezzoMax
          schema:
            type: number
            format: float
          description: Filtro per prezzo massimo del prodotto.
      responses:
        '200':
          description: Elenco dei prodotti restituito con successo.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prodotto'
    post:
      security:
        - bearerAuth: []
      summary: Aggiunge un nuovo prodotto al menu
      tags: [Catalogo]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Prodotto'
      responses:
        '201':
          description: Prodotto aggiunto con successo.
        '403':
          description: Accesso negato. Solo il proprietario può effettuare questa operazione.
        '500':
          description: Errore interno del server.
  /prodotti/{id}:
    put:
      security:
        - bearerAuth: []
      summary: Aggiorna un prodotto esistente
      tags: [Catalogo]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Prodotto'
      responses:
        '200':
          description: Prodotto aggiornato con successo.
        '403':
          description: Accesso negato.
        '500':
          description: Errore interno del server.
  /prodotti/{idProdotto}/caratteristiche:
    post:
      security:
        - bearerAuth: []
      summary: Assegna una caratteristica esistente a un prodotto.
      tags: [Catalogo]
      parameters:
        - in: path
          name: idProdotto
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Caratteristica'
      responses:
        '201':
          description: Caratteristica assegnata con successo.
        '400':
          description: ID caratteristica mancante.
        '403':
          description: Accesso negato.
        '500':
          description: Errore durante l'assegnazione.
  /prodotti/{idProdotto}/caratteristiche/{idCaratteristica}:
    delete:
      security:
        - bearerAuth: [] # Questo endpoint richiede autenticazione con token JWT
      summary: Rimuove una caratteristica da un prodotto.
      description: Rimuove una caratteristica specifica da un prodotto esistente.
      tags: [Catalogo]
      parameters:
        - in: path
          name: idProdotto
          required: true
          schema:
            type: integer
          description: ID del prodotto da cui rimuovere la caratteristica.
        - in: path
          name: idCaratteristica
          required: true
          schema:
            type: integer
          description: ID della caratteristica da rimuovere dal prodotto.
      responses:
        '204':
          description: Caratteristica rimossa con successo.
        '404':
          description: Prodotto o caratteristica non trovata.
  /prodotti/{idProdotto}/ingredienti:
    get:
      summary: Elenco ingredienti di un prodotto.
      description: Restituisce la lista degli ingredienti associati a un prodotto specifico. Se si utilizza un token di Proprietario o di Personale, vengono restituite anche le quantità per ciascun ingrediente.
      tags: [Catalogo]
      parameters:
        - in: path
          name: idProdotto
          required: true
          schema:
            type: integer
          description: ID del prodotto di cui ottenere gli ingredienti.
      responses:
        '200':
          description: Elenco degli ingredienti restituito con successo.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingrediente'
...
components:
  schemas:
...
    Caratteristica:
      type: object
      properties:
        idCaratteristica:
          type: integer
          example: 1
        nome:
          type: string
          example: "Piccante"
        differenzaPrezzo:
          type: number
          format: float
          example: -0.50
        isDefault:
          type: boolean
          example: false
        idGmc:
          type: integer
          example: 2
        gruppo:
          type: string
          description: Il nome del gruppo di caratteristiche in mutua esclusione se presente
          example: "Zucchero"
    #############################################
    Prodotto:
      type: object
      properties:
        idProdotto:
          type: integer
          example: 1
        categoria:
          type: string
          example: "Pizza"
        nome:
          type: string
          example: "Pizza Margherita"
        descrizione:
          type: string
          example: "Pizza con pomodoro, mozzarella e basilico."
        prezzoBase:
          type: number
          format: float
          example: 5.99
        immagine:
          type: string
          example: "http://localhost:8080/Fooody/images/pizza_margherita.jpg"
        tempoPreparazione:
          type: integer
          description: Tempo di preparazione stimato in minuti
          example: 15
        procedura:
          type: string
          description: Procedura di preparazione del prodotto necessaria alla cucina.
          example: "1. Stendere la pasta. 2. Aggiungere il pomodoro. 3. Aggiungere la mozzarella. 4. Cuocere in forno per 15 minuti."
        caratteristiche:
          type: array
          items:
            $ref: '#/components/schemas/Caratteristica'
    #############################################
    Ingrediente:
      type: object
      properties:
        idIngrediente:
          type: integer
          example: 1
        nome:
          type: string
          example: "Pomodoro"
        quantita:
          type: string
          example: "100g"
    #############################################
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Utilizza un token JWT per autenticarsi nelle richieste alle risorse protette. Il token deve essere incluso nell'header Authorization con il prefisso "Bearer".